# .archive/reference-snapshot/sih/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketResponse": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "availability_signal": {
            "title": "Availability Signal",
            "type": "object"
          },
          "bunker": {
            "title": "Bunker",
            "type": "object"
          },
          "charter_opportunity_score_generic": {
            "title": "Charter Opportunity Score Generic",
            "type": "object"
          },
          "commodities": {
            "title": "Commodities",
            "type": "object"
          },
          "fx": {
            "title": "Fx",
            "type": "object"
          },
          "indices": {
            "title": "Indices",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": {
              "enum": [
                "measured",
                "derived",
                "estimated",
                "simulated_demo",
                "expert_set"
              ],
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "as_of",
          "indices",
          "bunker",
          "commodities",
          "fx",
          "availability_signal",
          "charter_opportunity_score_generic",
          "provenance",
          "request_id"
        ],
        "title": "MarketResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for chartering bulk carriers into India's East Coast ports.\n\n**Forecast the market \u2192 understand the constraints \u2192 simulate charter strategies \u2192\nrecommend the best decision.**\n\nThis is not a freight price oracle. It is a constrained decision engine that treats\nthe freight forecast as an uncertain input, not as the answer. Every response carries\n`provenance`, `as_of` and `assumptions`, and every infeasible option comes back with\nthe constraint and the source that ruled it out.\n\nProvenance values: `measured` (from an official published record), `derived` (computed\nby one of our models from measured inputs), `estimated` (a documented judgement where no\npublication exists), `expert_set` (a hand-curated parameter), `simulated_demo` (a\nseeded demo series standing in for a feed that needs credentials).\n",
    "title": "SIH26006 \u2014 Charter Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/market": {
      "get": {
        "operationId": "market_api_v1_market_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market dashboard",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/backhaul": {
      "get": {
        "operationId": "backhaul_api_v1_market_backhaul_get",
        "parameters": [
          {
            "in": "query",
            "name": "origin_port",
            "required": true,
            "schema": {
              "title": "Origin Port",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destination_port",
            "required": true,
            "schema": {
              "title": "Destination Port",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vessel_class",
            "required": false,
            "schema": {
              "default": "Supramax",
              "title": "Vessel Class",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backhaul insight for a lane",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/ballast-matrix": {
      "get": {
        "operationId": "ballast_matrix_api_v1_market_ballast_matrix_get",
        "parameters": [
          {
            "in": "query",
            "name": "vessel_class",
            "required": false,
            "schema": {
              "default": "Supramax",
              "title": "Vessel Class",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_port",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Port"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ballast cost matrix and the premium owners price in",
        "tags": [
          "market"
        ]
      }
    },
    "/api/v1/market/low-demand": {
      "get": {
        "operationId": "low_demand_api_v1_market_low_demand_get",
        "parameters": [
          {
            "in": "query",
            "name": "vessel_class",
            "required": false,
            "schema": {
              "default": "Supramax",
              "title": "Vessel Class",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 12,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Six-month demand-regime strip for a vessel class",
        "tags": [
          "market"
        ]
      }
    },
...
  },
...
}